requirements.txt
wxPython>=4.0
pyserial
